+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Constants/PublicHookNames.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\GatoGraphQL\Constants;
+
+class PublicHookNames
+{
+    public final const BEFORE_APP_IS_LOADED_GRAPHQL_ENDPOINT_PATHS = 'before_app_is_loaded:graphql_endpoint_paths';
+}