+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Hooks/ApplicationPasswordAuthorizationHookSet.php
@@ -6,6 +6,7 @@ namespace GatoGraphQL\GatoGraphQL\Hooks;
 
 use GatoGraphQL\GatoGraphQL\App;
 use GatoGraphQL\GatoGraphQL\AppHelpers;
+use GatoGraphQL\GatoGraphQL\ContainerLess\BeforeAppIsLoadedStaticHelpers;
 use GatoGraphQL\GatoGraphQL\FeedbackItemProviders\ErrorFeedbackItemProvider;
 use GatoGraphQL\GatoGraphQL\Request\PrematureRequestServiceInterface;
 use PoP\ComponentModel\Engine\EngineHookNames;
@@ -64,6 +65,11 @@ class ApplicationPasswordAuthorizationHookSet extends AbstractHookSet
             'application_password_failed_authentication',
             $this->handleError(...)
         );
+
+        /**
+         * As we've now loaded the proper logic, we can remove the workaround hooks
+         */
+        BeforeAppIsLoadedStaticHelpers::removeApplicationPasswordHooks();
     }
 
     /**